move tests, remove obsolete snapshot
