+++ src/validation/tests/naming_validation_test.rs
@@ -49,46 +49,3 @@ assert_with_type_name!(WSTRING);
 assert_with_type_name!(CHAR);
 assert_with_type_name!(WCHAR);
 assert_with_type_name!(VOID);
-
-#[test]
-fn two_identical_vlas_in_same_pou_arent_duplicated_in_symbol_map() {
-    let diag = parse_and_validate(
-        r#"
-        FUNCTION foo : INT
-        VAR_INPUT{ref}
-            vla1 : ARRAY[ * , * ] OF INT;
-            vla2 : ARRAY[ * , * ] OF INT;
-        END_VAR
-        END_FUNCTION
-
-        FUNCTION bar : DINT
-        VAR_INPUT{ref}
-            vla1 : ARRAY[ * , * ] OF LINT;
-            vla2 : ARRAY[ * , * ] OF SINT;
-            vla3 : ARRAY[ * , *, * ] OF SINT;
-        END_VAR
-        END_FUNCTION
-    "#,
-    );
-
-    assert_validation_snapshot!(&diag);
-}
-
-#[test]
-fn global() {
-    let diag = parse_and_validate(
-        r#"
-        VAR_GLOBAL
-            vla : ARRAY[*, *] OF DINT;
-        END_VAR
-        
-        FUNCTION foo : DINT
-        VAR_IN_OUT
-            arr : ARRAY[*, *] OF DINT;
-        END_VAR
-        END_FUNCTION
-    "#,
-    );
-
-    assert_validation_snapshot!(&diag);
-}
