+++ src/validation/tests/variable_length_array_test.rs
@@ -178,3 +178,48 @@ mod assignment {
         assert_validation_snapshot!(diagnostics);
     }
 }
+
+mod naming {
+    use crate::{assert_validation_snapshot, test_utils::tests::parse_and_validate};
+    #[test]
+    fn two_identical_vlas_in_same_pou_arent_duplicated_in_symbol_map() {
+        let diag = parse_and_validate(
+            r#"
+        FUNCTION foo : INT
+        VAR_INPUT{ref}
+            vla1 : ARRAY[ * , * ] OF INT;
+            vla2 : ARRAY[ * , * ] OF INT;
+        END_VAR
+        END_FUNCTION
+
+        FUNCTION bar : DINT
+        VAR_INPUT{ref}
+            vla1 : ARRAY[ * , * ] OF LINT;
+            vla2 : ARRAY[ * , * ] OF SINT;
+            vla3 : ARRAY[ * , *, * ] OF SINT;
+        END_VAR
+        END_FUNCTION
+    "#,
+        );
+
+        assert_eq!(diag.len(), 0);
+    }
+
+    #[test]
+    fn global_vla_does_not_cause_name_conflict() {
+        let diag = parse_and_validate(
+            r#"
+        VAR_GLOBAL
+            vla : ARRAY[*, *] OF DINT;
+        END_VAR
+        
+        FUNCTION foo : DINT
+        VAR_IN_OUT
+            arr : ARRAY[*, *] OF DINT;
+        END_VAR
+        END_FUNCTION
+    "#,
+        );
+        assert_validation_snapshot!(diag);
+    }
+}
